CON-225: Disable possible nondeterminism.
We're using wasmi and pwasm-utils where inside their usage of data
structures is interesting - for `std` there is HashSet used, and
for lack of `std` BTreeSet is used. In other places `HashSet` is swapped
with `hashbrown` for non-`std` builds.

It is good idea to disable this behaviour, and exclude possible sources
of nondeterminism.