EE-637: Change serialization of arguments.
This commit removes `Vec<Vec<u8>>` intermediary step for deserialization
and serializes/deserializes vector of Values as ordinary vector. Former
approach serialized each of the values as `Vec<u8>` which contained
length prefix for the data, with new approach the safety guarantees
should be comparable as each `Value` variant serializes its data with an
ID prefix.