EE-906: Handle MapResized error when creating tx.
This will transparently retry begin_ro_tx/begin_rw_tx whenever a new
operation returns a "MapResized" error. This can happen when other
process resized the map, so we have to handle it by resizing the map to
0 so current process can pick up new size correctly.

This is called in a loop for safety, and from initial tests it appears
that the LMDB library internally waits (by returning MapResized error)
so this shouldn't be a busy waiting.